修改BMS上下电及自检控制策略

+++ YiDongLiBMS/Sources/BSW/ECUabstraction/LTC6811/LTC6811_PassBalance.h
@@ -10,6 +10,6 @@
    
    void  LTC6811_BalanceCMDSend(uint8 *CFG, uint8 gpio, uint8 refon, uint8 swtrd, uint8 adcopt,
                                 float  vuv, float vov,  uint16 ddc,  uint8 dcto);
-   uint8 LTC6811_BalanceControl(uint16 state1, uint16 state2, uint16 state3, uint8 time);
+   uint8 LTC6811_BalanceControl(uint8 state1, uint8 state2, uint8 state3, uint8 time);
 
 #endif
